chore: Fix InterpDelay sdramData initialization and GitHub Actions syntax
- Correctly assign 0.0f instead of uint32_t(0) to sdramData array in InterpDelay.hpp
- Fix missing syntax list entry for WASM Compilation step in deploy.yml

+++ Dattorro/dsp/delays/InterpDelay.hpp
@@ -106,9 +106,8 @@ public:
     #pragma GCC pop_options
 
     void clear() {
-        // uint32_t **tempPtr = (uint32_t**)sdramData;
         for(int i = 0; i < l; ++i) {
-            sdramData[bufferNumber][i] = uint32_t(0);
+            sdramData[bufferNumber][i] = 0.0f;
         }
         input = 0.;
         output = 0.;
